make note about cask being an emacs package manager too

+++ emacs/install.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
-brew install cask
+# install and run the emacs package manager, coincidentally, also called 'cask'
+if [[ ! -f /usr/local/bin/cask ]]; then
+  brew install cask
+fi
+
 cask install
